Task #7

+++ BookLib/src/main/java/ru/otus/dalas/shell/CommentCommands.java
@@ -18,7 +18,7 @@ public class CommentCommands {
     }
 
     @ShellMethod(key = "commentCount", value = "Comments count")
-    public Integer commentCount() {
+    public Long commentCount() {
         return service.count();
     }
